Update README.md
This project is an application that lists timetables for different cities around the world. The user types the name of the city and it is added to this list, showing the time in each location simultaneously. In companies with teams allocated in several countries, this type of functionality is very useful. The list of cities is obtained by the City Timezones library, which converts the name of the city typed by the user into the respective timezone. Each time corresponding to a timezone is obtained by the React Live Clock library. Some Bootstrap components for input field are also used